Move Dot to base package
