+++ src/main/scala/com/nthportal/versions/v3/Version.scala
@@ -1,8 +1,6 @@
 package com.nthportal.versions
 package v3
 
-import com.nthportal.versions.util.Dot
-
 case class Version(major: Int, minor: Int, patch: Int) extends Ordered[Version] {
   // Validate values
   require(major >= 0 && minor >= 0 && patch >= 0, "major, minor, and patch values must all be >= 0")
